+++ rrgen/tests/fixtures/test1/expected/before_all.txt
@@ -1,9 +1,9 @@
 // Both struct should be injected
 
+//fields
 pub struct World {
-  //fields
 }
 
+//fields
 pub struct World2 {
-  //fields
 }
